+++ pyneqsys/tests/test_pysym.py
@@ -16,6 +16,6 @@ def test_pysym_SymbolicSys_from_callback():
         Matrix=pysym.Matrix)
 
     x, sol = ss.solve('scipy', [1, 0], [3])
-    assert sol.success
+    assert sol['success']
     assert abs(x[0] - 0.8411639) < 2e-7
     assert abs(x[1] - 0.1588361) < 2e-7