+++ pyneqsys/tests/test_symbolic.py
@@ -15,7 +15,7 @@ def test_SymbolicSys():
         return 1/(x+a)**t + b
     neqsys = SymbolicSys([a, b], [f(0) - 1, f(1) - 0], [t])
     ab, sol = neqsys.solve_scipy([0.5, -0.5], 1)
-    assert sol.success
+    assert sol['success']
     assert abs(ab[0] - (-1/2 + 5**0.5/2)) < 1e-10
     assert abs(ab[1] - (1/2 - 5**0.5/2)) < 1e-10
 
@@ -23,7 +23,7 @@ def test_SymbolicSys():
 def test_symbolicsys__from_callback():
     ss = SymbolicSys.from_callback(f, 2, 1)
     x, sol = ss.solve_scipy([1, 0], [3])
-    assert sol.success
+    assert sol['success']
     assert abs(x[0] - 0.8411639) < 2e-7
     assert abs(x[1] - 0.1588361) < 2e-7
 
@@ -34,7 +34,7 @@ def test_symbolicsys__from_callback__no_params():
 
     ss = SymbolicSys.from_callback(_nf, 2)
     x, sol = ss.solve_scipy([.7, .3])
-    assert sol.success
+    assert sol['success']
     assert abs(x[0] - 0.8411639) < 2e-7
     assert abs(x[1] - 0.1588361) < 2e-7
 
@@ -42,7 +42,7 @@ def test_symbolicsys__from_callback__no_params():
 def test_TransformedSys__from_callback():
     ts = TransformedSys.from_callback(f, 2, 1, transf_cbs=(sp.exp, sp.log))
     x, sol = ts.solve('scipy', [1, .1], [3])
-    assert sol.success
+    assert sol['success']
     assert abs(x[0] - 0.8411639) < 2e-7
     assert abs(x[1] - 0.1588361) < 2e-7