+++ CHANGES.rst
@@ -1,7 +1,9 @@
-v0.2.2
+v0.3.0
 ======
 - Added ChainedNeqSys
 - In NeqSys.solve() arg "solver" may now be None -> $NEQSYS_SOLVER
+- NeqSys.plot_series changed signature
+- New methods: NeqSys.plot_series_residuals(_internal)
 
 v0.2.1
 ======